[refactor] Rename testing.py to _testing.py
